Extract icons.rs, and the OS integrations into platform.rs — Phase 2 done
The last Phase-2 item, and it is two things. icons.rs takes the favicon/logo
probing whole (sniff_mime, read_icon, probe_icon_dir, find_project_icon,
read_custom_icon + its test): content wins over extension, which is the one rule
that shapes it. platform.rs takes what was left that is OS integration and has
no other home — open_folder/reveal_path, the caffeinate/keep-awake cfg pair, and
the macOS legacy-localStorage recovery, with their three tests.

That makes platform.rs two halves and its module doc now says so. The leaves
import nothing from the crate, which is what lets every other module depend on
it; the integrations may, and do — set_caffeinate takes State<AppState>, so this
module now imports from the crate root at all. The first half must not.

lib.rs 1,097 -> 449, and its `mod tests` is gone: every test now has a module.
What it keeps is the Rust equivalent of Phase 1's "reduced to bootstrap" — the
mod list, AppState/Session, write_debug_file/log_frontend (the frontend's end of
the two-tier logging, beside install_panic_hook), confirm_quit, the tray mirror,
and run() with its invoke_handler! list.

Phase 2 totals: lib.rs 4,711 -> 449 (-90%), ten modules, 69 -> 77 tests.

A second verification trap, recorded in PLAN: **an appending slice must be
checked as a multiset** — every non-blank line of the merged file is one the
target already had or one the spec cut, with nothing missing. My append script's
header-skip used startsWith("//"), which also matches ///, so it silently ate
three lines of open_folder's doc comment. The multiset check caught it; nothing
else would have. Script fixed, merge redone, and the check now reports the
module-doc rewrite and the two new imports as the only differences.

Verification:
- icons.rs vs the pre-slice lib.rs, line-sorted: identical, 158 non-blank lines.
- platform.rs as a multiset: exact, 639 lines accounted for, before the two hand
  edits (the doc rewrite and `use tauri::State` / `use crate::AppState`).
- the remainder of lib.rs: 30 lines, each intended — mod icons;, the 6 icons::/
  platform:: qualifications, the import trims (base64 and State lost their last
  users), KeepAwake now imported from platform, and the empty test mod removed.

cargo test 77 passed; clean under the cfg flip. pnpm exec tsc --noEmit clean and
pnpm test 368 passed — unchanged, no TypeScript was touched in this phase.
