+++ src-tauri/src/lib.rs
@@ -9,6 +9,7 @@
 
 mod external;
 mod git;
+mod icons;
 mod platform;
 mod pty;
 mod tasks;
@@ -21,19 +22,15 @@ use std::collections::{HashMap, HashSet};
 use std::io::Write;
 use std::sync::Mutex;
 
-use base64::engine::general_purpose::STANDARD;
-use base64::Engine as _;
 use portable_pty::{ChildKiller, MasterPty};
 use tauri::menu::MenuBuilder;
 #[cfg(target_os = "macos")]
 use tauri::menu::{MenuItemBuilder, SubmenuBuilder};
 use tauri::tray::TrayIconBuilder;
-use tauri::{AppHandle, Emitter, Manager, State};
+use tauri::{AppHandle, Emitter, Manager};
 
-// Only the macOS legacy-localStorage reader still needs this; gated, or Windows
-// and Linux see an unused import.
-#[cfg(target_os = "macos")]
-use crate::platform::home_dir;
+#[cfg(windows)]
+use crate::platform::KeepAwake;
 use crate::telemetry::run_telemetry_server;
 
 pub(crate) struct Session {
@@ -72,115 +69,6 @@ pub(crate) struct AppState {
 }
 
 
-/// Open a project's folder in the OS file manager (Explorer / Finder / the
-/// desktop's default handler). Refuses a vanished directory rather than silently
-/// doing nothing — deleted worktrees are real.
-#[tauri::command]
-fn open_folder(dir: String) -> Result<(), String> {
-    if !std::path::Path::new(&dir).is_dir() {
-        return Err(format!("not a directory: {dir}"));
-    }
-    #[cfg(windows)]
-    {
-        // Explorer's shell parser only understands backslashes. Hand it a
-        // forward-slash path and it silently opens the user's Documents folder
-        // instead — no error, no clue. That's not hypothetical: an external
-        // session's project row carries the repo root from `git rev-parse`, and
-        // git emits `E:/Programming/…` on Windows. `Path::is_dir` accepts either
-        // form, so the guard above cannot catch it; normalize here.
-        let dir = dir.replace('/', "\\");
-        // explorer.exe is fire-and-forget: it hands off to the running shell and
-        // exits non-zero even when the window opened, so never wait on its status.
-        let sysroot = std::env::var("SystemRoot").unwrap_or_else(|_| r"C:\Windows".to_string());
-        std::process::Command::new(format!(r"{sysroot}\explorer.exe"))
-            .arg(&dir)
-            .spawn()
-            .map_err(|e| format!("open Explorer: {e}"))?;
-    }
-    #[cfg(target_os = "macos")]
-    {
-        std::process::Command::new("open")
-            .arg(&dir)
-            .spawn()
-            .map_err(|e| format!("open Finder: {e}"))?;
-    }
-    #[cfg(all(unix, not(target_os = "macos")))]
-    {
-        std::process::Command::new("xdg-open")
-            .arg(&dir)
-            .spawn()
-            .map_err(|e| format!("xdg-open: {e}"))?;
-    }
-    Ok(())
-}
-
-/// Reveal a file in the OS file manager, selected — the "↗ Reveal source" action on
-/// a task, which answers "where did this come from?". `dir` is the project root and
-/// `rel` the repo-relative source file (`.vscode/tasks.json`); revealing by relative
-/// path keeps a `..` from a malformed override from escaping the project. Falls back
-/// to opening the containing folder where the OS can't select a file.
-#[tauri::command]
-fn reveal_path(dir: String, rel: String) -> Result<(), String> {
-    let root = std::path::Path::new(&dir);
-    if !root.is_dir() {
-        return Err(format!("not a directory: {dir}"));
-    }
-    // Reject anything that would climb out of the project — `rel` is nominally
-    // repo-relative, but it reaches us from discovery data, so don't trust it. A
-    // `..` climbs out; a rooted (`/x`) or drive-prefixed (`C:x`) component is what
-    // `join` replaces the whole base with — note `/x` is *not* `is_absolute()` on
-    // Windows (no prefix), so the component check, not `is_absolute`, is the guard.
-    let rel_path = std::path::Path::new(&rel);
-    if rel_path.is_absolute()
-        || rel_path.components().any(|c| {
-            matches!(
-                c,
-                std::path::Component::ParentDir
-                    | std::path::Component::RootDir
-                    | std::path::Component::Prefix(_)
-            )
-        })
-    {
-        return Err(format!("not a project-relative path: {rel}"));
-    }
-    let target = root.join(rel_path);
-    // A source that no longer exists (a task from a file since deleted): reveal the
-    // project folder rather than erroring on a path the user can't do anything about.
-    let exists = target.is_file();
-    #[cfg(target_os = "macos")]
-    {
-        let mut c = std::process::Command::new("open");
-        if exists {
-            c.arg("-R").arg(&target);
-        } else {
-            c.arg(root);
-        }
-        c.spawn().map_err(|e| format!("open Finder: {e}"))?;
-    }
-    #[cfg(windows)]
-    {
-        let sysroot = std::env::var("SystemRoot").unwrap_or_else(|_| r"C:\Windows".to_string());
-        let mut c = std::process::Command::new(format!(r"{sysroot}\explorer.exe"));
-        if exists {
-            // /select, takes one argument: the file to highlight in its folder.
-            c.arg(format!("/select,{}", target.display().to_string().replace('/', "\\")));
-        } else {
-            c.arg(dir.replace('/', "\\"));
-        }
-        c.spawn().map_err(|e| format!("open Explorer: {e}"))?;
-    }
-    #[cfg(all(unix, not(target_os = "macos")))]
-    {
-        // No portable "select the file" across Linux file managers; open the folder.
-        let open_target = if exists { target.parent().unwrap_or(root) } else { root };
-        std::process::Command::new("xdg-open")
-            .arg(open_target)
-            .spawn()
-            .map_err(|e| format!("xdg-open: {e}"))?;
-    }
-    Ok(())
-}
-
 /// Persist a debug snapshot (JSON built by the frontend) to a fixed, discoverable
 /// path so an external tool — or an LLM agent debugging the running app — can read
 /// live state and the recent event log. Returns the path written.
@@ -209,435 +97,7 @@ fn log_frontend(level: String, msg: String) {
     }
 }
 
-/// A caffeinate flag we're willing to pass through: a short-option cluster over
-/// the sleep-assertion letters (`-d -i -m -s -u`, or combined like `-dimsu`),
-/// the `-t` timeout switch, or a bare decimal number (its seconds argument).
-/// Everything the UI sends is a fixed preset, so this is just belt-and-braces
-/// against ever handing an arbitrary string to the shell-less spawn.
-#[cfg(not(windows))]
-fn valid_caffeinate_flag(f: &str) -> bool {
-    if let Some(rest) = f.strip_prefix('-') {
-        return !rest.is_empty() && rest.chars().all(|c| "dimsut".contains(c));
-    }
-    !f.is_empty() && f.chars().all(|c| c.is_ascii_digit())
-}
 
-/// Toggle a macOS `caffeinate` power-assertion on or off. Only ever one child
-/// runs: any existing one is killed first, so switching presets is just a
-/// stop+restart. `active=false` (or an empty `flags`) simply stops it.
-#[cfg(not(windows))]
-#[tauri::command]
-fn set_caffeinate(state: State<AppState>, active: bool, flags: Vec<String>) -> Result<(), String> {
-    let mut guard = state.caffeinate.lock().unwrap();
-    // Tear down whatever's running (also reaps a child that self-exited on -t).
-    if let Some(mut child) = guard.take() {
-        let _ = child.kill();
-        let _ = child.wait();
-    }
-    if !active || flags.is_empty() {
-        return Ok(());
-    }
-    if let Some(bad) = flags.iter().find(|f| !valid_caffeinate_flag(f)) {
-        return Err(format!("refusing unknown caffeinate flag: {bad}"));
-    }
-    // `-w <our pid>`: caffeinate exits on its own the moment Episko does, so a
-    // crash or force-quit can't leave the display pinned awake.
-    let mut cmd = std::process::Command::new("/usr/bin/caffeinate");
-    cmd.arg("-w").arg(std::process::id().to_string());
-    for f in &flags {
-        cmd.arg(f);
-    }
-    cmd.stdin(std::process::Stdio::null())
-        .stdout(std::process::Stdio::null())
-        .stderr(std::process::Stdio::null());
-    let child = cmd.spawn().map_err(|e| format!("caffeinate: {e}"))?;
-    *guard = Some(child);
-    Ok(())
-}
-
-/// A live Windows power assertion. `SetThreadExecutionState` is scoped to the
-/// *calling thread* — the assertion dies with that thread — so we park a thread
-/// for exactly as long as the user wants the machine awake. That thread-scoping
-/// is also the safety net the macOS side gets from `caffeinate -w <our pid>`: a
-/// panic or a hard exit can't leave a Windows box pinned awake, because the
-/// thread goes with the process.
-#[cfg(windows)]
-struct KeepAwake {
-    /// Dropping this releases the assertion: the parked thread's `recv()` fails,
-    /// it clears the execution state and exits.
-    stop: Option<std::sync::mpsc::Sender<()>>,
-    thread: Option<std::thread::JoinHandle<()>>,
-}
-
-#[cfg(windows)]
-impl Drop for KeepAwake {
-    fn drop(&mut self) {
-        drop(self.stop.take());
-        // Join so the state is provably cleared before a replacement assertion
-        // is set up — otherwise a preset switch could race the old thread's
-        // clearing call and land on ES_CONTINUOUS (i.e. silently off).
-        if let Some(t) = self.thread.take() {
-            let _ = t.join();
-        }
-    }
-}
-
-/// Translate the macOS `caffeinate` flags the UI speaks into Windows execution
-/// state bits, so the frontend keeps one vocabulary for both platforms.
-///
-///   `-d` (display) → `ES_DISPLAY_REQUIRED`   `-i` / `-s` (idle/system) → `ES_SYSTEM_REQUIRED`
-///   `-m` (disk) and `-u` (user active) have no Windows equivalent and are dropped.
-///   `-t <sec>` is dropped too — the frontend's own timer disarms the preset.
-///
-/// Anything that asks for the display also implies the system, matching what a
-/// user means by "keep the screen on". Returns 0 when nothing was requested.
-///
-/// Deliberately *not* mapped: `ES_AWAYMODE_REQUIRED`. It's only honoured where
-/// the power policy enables away mode, and where it isn't the whole call fails
-/// (returns 0) — so asking for it would silently assert nothing at all.
-#[cfg(windows)]
-fn execution_state_for(flags: &[String]) -> u32 {
-    use windows_sys::Win32::System::Power::{ES_DISPLAY_REQUIRED, ES_SYSTEM_REQUIRED};
-    let mut es = 0u32;
-    for f in flags {
-        let Some(rest) = f.strip_prefix('-') else { continue }; // bare `-t` argument
-        for c in rest.chars() {
-            match c {
-                'd' => es |= ES_DISPLAY_REQUIRED | ES_SYSTEM_REQUIRED,
-                'i' | 's' => es |= ES_SYSTEM_REQUIRED,
-                _ => {} // m / u / t — nothing to assert
-            }
-        }
-    }
-    es
-}
-
-/// Toggle a Windows power assertion on or off — the `caffeinate` counterpart.
-/// Only ever one assertion is live: an existing one is dropped (which joins its
-/// thread and clears the state) first, so switching presets is a stop+restart.
-#[cfg(windows)]
-#[tauri::command]
-fn set_caffeinate(state: State<AppState>, active: bool, flags: Vec<String>) -> Result<(), String> {
-    use windows_sys::Win32::System::Power::{SetThreadExecutionState, ES_CONTINUOUS};
-    let mut guard = state.caffeinate.lock().unwrap();
-    guard.take(); // drop → releases whatever was asserted
-    if !active || flags.is_empty() {
-        return Ok(());
-    }
-    let es = execution_state_for(&flags);
-    if es == 0 {
-        return Err(format!("no Windows keep-awake equivalent for: {}", flags.join(" ")));
-    }
-    let (stop, rx) = std::sync::mpsc::channel::<()>();
-    // The assertion must be set *and* released on the same thread, so the whole
-    // lifetime lives inside this closure: assert, park, clear.
-    let (ready, ready_rx) = std::sync::mpsc::channel::<Result<(), String>>();
-    let thread = std::thread::spawn(move || {
-        // SAFETY: a plain flags-in/flags-out Win32 call with no pointers.
-        let prev = unsafe { SetThreadExecutionState(ES_CONTINUOUS | es) };
-        if prev == 0 {
-            let _ = ready.send(Err("SetThreadExecutionState refused the request".into()));
-            return;
-        }
-        let _ = ready.send(Ok(()));
-        let _ = rx.recv(); // park until the sender is dropped
-        // SAFETY: same call; ES_CONTINUOUS alone clears our assertion.
-        unsafe { SetThreadExecutionState(ES_CONTINUOUS) };
-    });
-    // Surface a refusal as a command error instead of a thread that quietly did
-    // nothing — the UI would otherwise paint the cup lit over a sleeping PC.
-    match ready_rx.recv() {
-        Ok(Ok(())) => {}
-        Ok(Err(e)) => {
-            let _ = thread.join();
-            return Err(e);
-        }
-        Err(_) => return Err("keep-awake thread died before asserting".into()),
-    }
-    *guard = Some(KeepAwake { stop: Some(stop), thread: Some(thread) });
-    Ok(())
-}
-
-// ---------- project favicon / logo discovery ----------
-
-#[derive(serde::Serialize)]
-struct ProjectIcon {
-    path: String,
-    data_uri: String,
-}
-
-/// Pick an image MIME from magic bytes, falling back to the file extension.
-/// Repos routinely ship a PNG named `favicon.ico`; trusting the extension would
-/// emit a `data:image/x-icon` URI wrapping PNG bytes, which the WebKit webview can
-/// refuse to render — so the icon would be "found" yet show as broken.
-fn sniff_mime(bytes: &[u8], ext: &str) -> Option<&'static str> {
-    if bytes.starts_with(b"\x89PNG\r\n\x1a\n") {
-        return Some("image/png");
-    }
-    if bytes.starts_with(&[0xFF, 0xD8, 0xFF]) {
-        return Some("image/jpeg");
-    }
-    if bytes.starts_with(b"GIF8") {
-        return Some("image/gif");
-    }
-    if bytes.len() >= 12 && &bytes[0..4] == b"RIFF" && &bytes[8..12] == b"WEBP" {
-        return Some("image/webp");
-    }
-    if bytes.starts_with(&[0x00, 0x00, 0x01, 0x00]) {
-        return Some("image/x-icon");
-    }
-    // SVG is text — look for a `<svg` tag near the start (past any XML prolog).
-    let head = &bytes[..bytes.len().min(256)];
-    if head.windows(4).any(|w| w.eq_ignore_ascii_case(b"<svg")) {
-        return Some("image/svg+xml");
-    }
-    // Couldn't sniff (e.g. an SVG with a long prolog) — trust the extension.
-    match ext {
-        "png" => Some("image/png"),
-        "svg" => Some("image/svg+xml"),
-        "ico" => Some("image/x-icon"),
-        "jpg" | "jpeg" => Some("image/jpeg"),
-        "webp" => Some("image/webp"),
-        "gif" => Some("image/gif"),
-        _ => None,
-    }
-}
-
-/// Read a candidate icon file into a base64 data-URI (small files only). The MIME
-/// is sniffed from content (see `sniff_mime`), not assumed from the extension.
-fn read_icon(p: &std::path::Path) -> Option<ProjectIcon> {
-    let meta = std::fs::metadata(p).ok()?;
-    if !meta.is_file() || meta.len() == 0 || meta.len() > 512 * 1024 {
-        return None;
-    }
-    let ext = p
-        .extension()
-        .and_then(|e| e.to_str())
-        .unwrap_or("")
-        .to_ascii_lowercase();
-    let bytes = std::fs::read(p).ok()?;
-    let mime = sniff_mime(&bytes, &ext)?;
-    let b64 = STANDARD.encode(&bytes);
-    Some(ProjectIcon {
-        path: p.to_string_lossy().to_string(),
-        data_uri: format!("data:{mime};base64,{b64}"),
-    })
-}
-
-/// Conventional favicon / logo spots relative to a web / Tauri / Electron project
-/// root. Returns the first that exists (no recursive walk — this stays cheap).
-fn probe_icon_dir(base: &std::path::Path) -> Option<ProjectIcon> {
-    const CANDIDATES: &[&str] = &[
-        "favicon.ico", "favicon.svg", "favicon.png",
-        "public/favicon.ico", "public/favicon.svg", "public/favicon.png",
-        "public/apple-touch-icon.png", "public/logo.svg", "public/logo.png",
-        "public/icon.svg", "public/icon.png",
-        "static/favicon.ico", "static/favicon.svg", "static/favicon.png",
-        "static/logo.svg", "static/logo.png",
-        "app/favicon.ico", "app/icon.png", "app/icon.svg",
-        "src/favicon.ico", "src/favicon.svg",
-        "src/assets/favicon.ico", "src/assets/favicon.svg", "src/assets/favicon.png",
-        "src/assets/logo.svg", "src/assets/logo.png",
-        "src/assets/icon.svg", "src/assets/icon.png",
-        "assets/favicon.png", "assets/logo.png", "assets/logo.svg", "assets/icon.png",
-        "resources/icon.png", "build/icon.png",
-        "src-tauri/icons/128x128.png", "src-tauri/icons/icon.png",
-    ];
-    CANDIDATES.iter().find_map(|rel| read_icon(&base.join(rel)))
-}
-
-/// Scour a project directory for a favicon / logo we can show as its sidebar
-/// glyph. Checks the conventional spots at the repo root, then — for monorepos
-/// that keep the web app in a subdirectory (e.g. `01_frontend/`, `frontend/`,
-/// `apps/web`) — one shallow level of subdirs, frontend-ish names first. This
-/// finds a nested `01_frontend/public/favicon.ico` without a deep filesystem walk.
-#[tauri::command]
-fn find_project_icon(dir: String) -> Option<ProjectIcon> {
-    let base = std::path::Path::new(&dir);
-    if !base.is_dir() {
-        return None;
-    }
-    // Fast path: conventional spots at the repo root.
-    if let Some(hit) = probe_icon_dir(base) {
-        return Some(hit);
-    }
-    // Fallback: probe immediate subdirectories, skipping heavy / build output dirs.
-    let mut subs: Vec<std::path::PathBuf> = std::fs::read_dir(base)
-        .ok()?
-        .filter_map(|e| e.ok())
-        .map(|e| e.path())
-        .filter(|p| p.is_dir())
-        .filter(|p| {
-            let name = p.file_name().and_then(|n| n.to_str()).unwrap_or("");
-            !name.starts_with('.')
-                && !matches!(
-                    name,
-                    "node_modules" | "target" | "dist" | "build" | "out"
-                        | "vendor" | "coverage" | "tmp" | "__pycache__"
-                )
-        })
-        .collect();
-    // Prefer frontend-ish directories, then fall back to alphabetical order so the
-    // choice is deterministic (e.g. `01_frontend` before `02_backend`).
-    subs.sort_by_key(|p| {
-        let name = p
-            .file_name()
-            .and_then(|n| n.to_str())
-            .unwrap_or("")
-            .to_ascii_lowercase();
-        let frontendish = ["front", "web", "client", "app", "ui", "site", "www"]
-            .iter()
-            .any(|k| name.contains(k));
-        (!frontendish, name)
-    });
-    subs.iter().find_map(|p| probe_icon_dir(p))
-}
-
-/// Load a user-picked image as a project's logo. Deliberately runs the same
-/// sniff/size gate as discovery (`read_icon`), so a file that isn't really an
-/// image — or one too big to sit in localStorage as a data-URI — is rejected here
-/// instead of becoming a broken `<img>` in the sidebar.
-#[tauri::command]
-fn read_custom_icon(path: String) -> Result<ProjectIcon, String> {
-    read_icon(std::path::Path::new(&path))
-        .ok_or_else(|| "Not a usable image (PNG, SVG, ICO, JPEG, WEBP or GIF, max 512 KB)".to_string())
-}
-
-// ---------- one-time recovery of localStorage stranded by a rename ----------
-//
-// macOS keys a WKWebView app's localStorage to its bundle identifier, so the
-// rename `io.respeak.cclauncher` (Muster) -> `io.respeak.episko` pointed the app
-// at a fresh, empty store and left every `cc-*` key (the daily cost rollup, the
-// project/session roster, favourites, colours, icons) behind under the old id.
-// This reads those keys straight off the old on-disk SQLite so the frontend can
-// import any it doesn't already have (fill-absent-only — it never clobbers data
-// this install legitimately holds; see the boot shim in `main.ts`). Returns an
-// empty map wherever there is nothing to recover, so the caller is a no-op.
-
-/// WebKit stores localStorage values as UTF-16LE BLOBs (no BOM).
-#[cfg(target_os = "macos")]
-fn decode_utf16le(bytes: &[u8]) -> String {
-    let units: Vec<u16> = bytes
-        .chunks_exact(2)
-        .map(|c| u16::from_le_bytes([c[0], c[1]]))
-        .collect();
-    String::from_utf16_lossy(&units)
-}
-
-/// Depth-limited walk for `localstorage.sqlite3` under a WebsiteData/Default
-/// tree (the origin salt/hash dirs are minted at runtime, so the path isn't
-/// known ahead of time). Returns the largest match — the store with the most
-/// data — since a stale empty one can linger beside the real one.
-#[cfg(target_os = "macos")]
-fn newest_localstorage_db(base: &std::path::Path) -> Option<std::path::PathBuf> {
-    let mut best: Option<(u64, std::path::PathBuf)> = None;
-    let mut stack = vec![(base.to_path_buf(), 0u8)];
-    while let Some((dir, depth)) = stack.pop() {
-        if depth > 6 {
-            continue;
-        }
-        let Ok(entries) = std::fs::read_dir(&dir) else {
-            continue;
-        };
-        for e in entries.flatten() {
-            let p = e.path();
-            if p.is_dir() {
-                stack.push((p, depth + 1));
-            } else if p.file_name().and_then(|n| n.to_str()) == Some("localstorage.sqlite3") {
-                let sz = e.metadata().map(|m| m.len()).unwrap_or(0);
-                if best.as_ref().map_or(true, |(b, _)| sz > *b) {
-                    best = Some((sz, p));
-                }
-            }
-        }
-    }
-    best.map(|(_, p)| p)
-}
-
-/// Read every `cc-*` key from a WebKit `localstorage.sqlite3`. Works on a private
-/// copy (never the original) so SQLite can fold in a stranded `-wal` — the old
-/// store may hold uncheckpointed writes the main file doesn't yet have.
-#[cfg(target_os = "macos")]
-fn read_itemtable_cc_keys(
-    db: &std::path::Path,
-) -> Result<std::collections::HashMap<String, String>, String> {
-    let tmp_dir = std::env::temp_dir().join("cc-launcher");
-    let _ = std::fs::create_dir_all(&tmp_dir);
-    let tmp = tmp_dir.join("legacy-localstorage.sqlite3");
-    for ext in ["sqlite3", "sqlite3-wal", "sqlite3-shm"] {
-        let _ = std::fs::remove_file(tmp.with_extension(ext));
-    }
-    std::fs::copy(db, &tmp).map_err(|e| format!("copy legacy store: {e}"))?;
-    for ext in ["sqlite3-wal", "sqlite3-shm"] {
-        let src = db.with_extension(ext);
-        if src.exists() {
-            let _ = std::fs::copy(&src, tmp.with_extension(ext));
-        }
-    }
-
-    let read = (|| -> Result<std::collections::HashMap<String, String>, String> {
-        let conn = rusqlite::Connection::open(&tmp).map_err(|e| format!("open legacy store: {e}"))?;
-        let mut stmt = conn
-            .prepare("SELECT key, value FROM ItemTable")
-            .map_err(|e| format!("query legacy store: {e}"))?;
-        let rows = stmt
-            .query_map([], |row| {
-                let key: String = row.get(0)?;
-                let val: rusqlite::types::Value = row.get(1)?;
-                Ok((key, val))
-            })
-            .map_err(|e| format!("read legacy rows: {e}"))?;
-        let mut out = std::collections::HashMap::new();
-        for (key, val) in rows.flatten() {
-            if !key.starts_with("cc-") {
-                continue;
-            }
-            let text = match val {
-                rusqlite::types::Value::Blob(b) => decode_utf16le(&b),
-                rusqlite::types::Value::Text(t) => t,
-                _ => continue,
-            };
-            out.insert(key, text);
-        }
-        Ok(out)
-    })();
-
-    for ext in ["sqlite3", "sqlite3-wal", "sqlite3-shm"] {
-        let _ = std::fs::remove_file(tmp.with_extension(ext));
-    }
-    read
-}
-
-#[tauri::command]
-fn read_legacy_localstorage() -> Result<std::collections::HashMap<String, String>, String> {
-    #[cfg(target_os = "macos")]
-    {
-        // Previous bundle ids, newest-first. Only the installed-app id belongs
-        // here; the `pnpm tauri dev` store ("muster") is intentionally excluded.
-        const LEGACY_IDS: &[&str] = &["io.respeak.cclauncher"];
-        let home = home_dir();
-        for id in LEGACY_IDS {
-            let base = std::path::Path::new(&home)
-                .join("Library/WebKit")
-                .join(id)
-                .join("WebsiteData/Default");
-            let Some(db) = newest_localstorage_db(&base) else {
-                continue;
-            };
-            let map = read_itemtable_cc_keys(&db)?;
-            if !map.is_empty() {
-                return Ok(map);
-            }
-        }
-        Ok(std::collections::HashMap::new())
-    }
-    #[cfg(not(target_os = "macos"))]
-    {
-        // Windows/WebView2 keeps localStorage in a LevelDB, not SQLite, and the
-        // pre-rename Windows build predates this scheme — nothing to recover.
-        Ok(std::collections::HashMap::new())
-    }
-}
 
 // ---------- app quit ----------
 
@@ -935,7 +395,7 @@ pub fn run() {
             git::git_action,
             pty::session_resources,
             git::create_worktree,
-            set_caffeinate,
+            platform::set_caffeinate,
             telemetry::resolve_permission,
             git::list_worktrees,
             git::remove_worktree,
@@ -962,11 +422,11 @@ pub fn run() {
             usage::read_transcript,
             usage::list_past_sessions,
             usage::token_usage_by_day,
-            find_project_icon,
-            read_custom_icon,
-            read_legacy_localstorage,
-            open_folder,
-            reveal_path,
+            icons::find_project_icon,
+            icons::read_custom_icon,
+            platform::read_legacy_localstorage,
+            platform::open_folder,
+            platform::reveal_path,
             write_debug_file,
             log_frontend,
             update_tray,
@@ -986,112 +446,4 @@ pub fn run() {
             tauri::RunEvent::Exit => log::info!("exit · clean shutdown"),
             _ => {}
         });
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    // Ditto: the only test left in this file is the macOS-only legacy-store one.
-    #[cfg(target_os = "macos")]
-    use crate::testutil::scratch_dir;
-
-    /// The legacy-store reader must decode WebKit's UTF-16LE BLOBs, keep only
-    /// `cc-*` keys, and survive a stranded `-wal`. Fixture mirrors the real
-    /// `ItemTable` schema so no machine-specific store is needed.
-    #[cfg(target_os = "macos")]
-    #[test]
-    fn read_itemtable_decodes_utf16le_and_filters_cc_keys() {
-        let dir = std::env::temp_dir().join(format!(
-            "episko-legacy-test-{}-{}",
-            std::process::id(),
-            COUNTER.fetch_add(1, Ordering::SeqCst)
-        ));
-        std::fs::create_dir_all(&dir).unwrap();
-        let db = dir.join("localstorage.sqlite3");
-        let _ = std::fs::remove_file(&db);
-        let utf16le = |s: &str| -> Vec<u8> { s.encode_utf16().flat_map(|u| u.to_le_bytes()).collect() };
-        {
-            let c = rusqlite::Connection::open(&db).unwrap();
-            c.execute_batch(
-                "CREATE TABLE ItemTable (key TEXT UNIQUE ON CONFLICT REPLACE, value BLOB NOT NULL ON CONFLICT FAIL);",
-            )
-            .unwrap();
-            c.execute(
-                "INSERT INTO ItemTable(key,value) VALUES(?1,?2)",
-                rusqlite::params!["cc-usage", utf16le(r#"{"2026-07-21":603.45}"#)],
-            )
-            .unwrap();
-            c.execute(
-                "INSERT INTO ItemTable(key,value) VALUES(?1,?2)",
-                rusqlite::params!["not-a-cc-key", utf16le("ignore me")],
-            )
-            .unwrap();
-        }
-        let map = read_itemtable_cc_keys(&db).unwrap();
-        assert_eq!(map.get("cc-usage").map(String::as_str), Some(r#"{"2026-07-21":603.45}"#));
-        assert!(!map.contains_key("not-a-cc-key"));
-        let _ = std::fs::remove_dir_all(&dir);
-    }
-
-
-    /// The one piece of the Windows keep-awake path that isn't a Win32 call: the
-    /// translation of the UI's `caffeinate` flags into execution-state bits.
-    #[cfg(windows)]
-    #[test]
-    fn caffeinate_flags_map_to_execution_state() {
-        use windows_sys::Win32::System::Power::{ES_DISPLAY_REQUIRED, ES_SYSTEM_REQUIRED};
-        let f = |args: &[&str]| execution_state_for(&args.iter().map(|s| s.to_string()).collect::<Vec<_>>());
-
-        // Asking for the display implies the system stays powered too.
-        assert_eq!(f(&["-d"]), ES_DISPLAY_REQUIRED | ES_SYSTEM_REQUIRED);
-        assert_eq!(f(&["-i"]), ES_SYSTEM_REQUIRED);
-        assert_eq!(f(&["-dimsu"]), ES_DISPLAY_REQUIRED | ES_SYSTEM_REQUIRED);
-        // The timer preset: `-t` and its bare seconds argument assert nothing on
-        // their own, and must not be mistaken for a flag cluster.
-        assert_eq!(f(&["-di", "-t", "3600"]), ES_DISPLAY_REQUIRED | ES_SYSTEM_REQUIRED);
-        // Nothing translatable → 0, which the command reports as an error rather
-        // than lighting the cup over a machine that will happily sleep.
-        assert_eq!(f(&["-m"]), 0);
-        assert_eq!(f(&[]), 0);
-    }
-
-    /// The macOS half of the same setting. `set_caffeinate` spawns without a shell,
-    /// but the flags still arrive from the frontend, so anything that isn't a
-    /// sleep-assertion cluster (or `-t`'s bare seconds) is refused rather than passed
-    /// to `/usr/bin/caffeinate`.
-    #[cfg(not(windows))]
-    #[test]
-    fn caffeinate_flags_are_whitelisted() {
-        for ok in ["-d", "-i", "-dimsu", "-t", "3600", "0"] {
-            assert!(valid_caffeinate_flag(ok), "{ok} is one of the UI's own presets");
-        }
-        for bad in ["", "-", "--", "-x", "-d;", "-d -i", "36 00", "/usr/bin/evil"] {
-            assert!(!valid_caffeinate_flag(bad), "{bad} must be refused");
-        }
-    }
-
-    /// Repos routinely ship a PNG named `favicon.ico`. Trusting the extension would
-    /// emit `data:image/x-icon` wrapping PNG bytes, which the webview may refuse —
-    /// the icon reads as "found" but renders broken. So content wins over extension,
-    /// and the extension is only the fallback.
-    #[test]
-    fn sniff_mime_trusts_content_over_extension() {
-        assert_eq!(sniff_mime(b"\x89PNG\r\n\x1a\nIHDR", "ico"), Some("image/png"));
-        assert_eq!(sniff_mime(&[0x00, 0x00, 0x01, 0x00, 0x01, 0x00], "png"), Some("image/x-icon"));
-        assert_eq!(sniff_mime(&[0xFF, 0xD8, 0xFF, 0xE0], "png"), Some("image/jpeg"));
-        assert_eq!(sniff_mime(b"GIF89a\x10\x00", "png"), Some("image/gif"));
-        assert_eq!(sniff_mime(b"RIFF\x00\x00\x00\x00WEBPVP8 ", "png"), Some("image/webp"));
-        assert_eq!(sniff_mime(b"<svg xmlns=\"http://www.w3.org/2000/svg\">", "png"), Some("image/svg+xml"));
-        // SVG is text, so it's found by tag — past an XML prolog, and case-insensitively.
-        assert_eq!(sniff_mime(b"<?xml version=\"1.0\"?>\n<SVG width=\"16\">", "bin"), Some("image/svg+xml"));
-        // Unsniffable (e.g. an SVG behind a long prolog) falls back to the extension.
-        assert_eq!(sniff_mime(b"", "svg"), Some("image/svg+xml"));
-        assert_eq!(sniff_mime(b"nothing recognisable", "webp"), Some("image/webp"));
-        // Neither content nor extension says image → no icon, rather than a guess.
-        assert_eq!(sniff_mime(b"nothing recognisable", "txt"), None);
-    }
-
-
-
-
 }
